Move ChangeLog constants into ChangeLog folder for easier access.
