+++ src/data-publication/ChangeLog/FilterBar.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react'
-import { PRODUCT_NAMES } from '../constants/publication-changes'
+import { PRODUCT_NAMES } from './constants'
 import { FilterResetButton } from './FilterResetButton'
 import spyGlass from '../../common/images/cf-gov-search.svg'
 import './FilterBar.css'
